fix(observability): split sentry privacy along its scalar seam; pin http_route rows
privacy.py sat one line under the 600-line cap; admitting http_route pushed
it over. Split along the seam the module docstring already named: generic
text scrubbers + closed scalar validators move to scalars.py (imported
directly by client.py, the package init, and request_telemetry — no
re-export), catalogs + projection stay in privacy.py. The http_route
validator matrix lives in its own small test file; the three integration
tests that pinned the old drop now pin a raw-uuid path dropping and the
template surviving.
