+++ server/tests/unit/test_sentry_integration.py
@@ -205,7 +205,7 @@ def test_set_server_sentry_tag_admits_only_catalog_rows(fake_sdk: _FakeSentrySdk
     sentry_integration.set_server_sentry_tag("http_route", "/orgs/{org_id}")
     sentry_integration.set_server_sentry_tag("session_id", "not-a-uuid")
     sentry_integration.set_server_sentry_tag("unknown_tag", "value")
-    assert fake_sdk.tag_calls == [("domain", "billing")]
+    assert fake_sdk.tag_calls == [("domain", "billing"), ("http_route", "/orgs/{org_id}")]
 
     fake_sdk.tag_calls.clear()
     session_id = str(uuid4())
@@ -222,7 +222,7 @@ def test_correlation_context_skips_unknown_and_invalid_entries(
             "user_id": "not-a-uuid",
             "session_id": "arbitrary",
             "worker_id": "arbitrary",
-            "http_route": "/raw/path",
+            "http_route": f"/orgs/{uuid4()}",
         }
     )
     assert fake_sdk.tag_calls == [("organization_id", organization_id)]
@@ -249,7 +249,7 @@ def test_capture_validates_scope_fields_and_ignores_fingerprint(
     sentry_integration.capture_server_sentry_exception(
         RuntimeError("boom"),
         level="warning",
-        tags={"domain": "billing", "http_route": "/raw", "unknown": "x"},
+        tags={"domain": "billing", "http_route": f"/orgs/{uuid4()}", "unknown": "x"},
         extras={
             "subject_id": subject_id,
             "drop_reason": "unhandled_event_type",
